+++ crates/wdl-lint/src/v1.rs
@@ -7,6 +7,7 @@ use wdl_ast::experimental::Diagnostics;
 
 use crate::TagSet;
 
+mod command_mixed_indentation;
 mod double_quotes;
 mod ending_newline;
 mod matching_parameter_meta;
@@ -17,6 +18,7 @@ mod preamble_whitespace;
 mod snake_case;
 mod whitespace;
 
+pub use command_mixed_indentation::*;
 pub use double_quotes::*;
 pub use ending_newline::*;
 pub use matching_parameter_meta::*;
@@ -67,6 +69,7 @@ pub fn rules() -> Vec<Box<dyn Rule>> {
         Box::new(PreambleCommentsRule),
         Box::new(MatchingParameterMetaRule),
         Box::new(WhitespaceRule),
+        Box::new(CommandSectionMixedIndentationRule),
     ];
 
     // Ensure all the rule ids are unique and pascal case